Extract TensorPlaceholder to a common interface
